+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,5 @@
+package duke.commands;
+
 public abstract class Command {
 
     public abstract void action();
