+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.commands;
+
+import duke.exceptions.InvalidCmdValueException;
+import duke.tasks.Task;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
+import duke.utils.Storage;
+
 import java.io.File;
 
 public class DeleteCommand extends Command {
@@ -9,7 +17,7 @@ public class DeleteCommand extends Command {
 
     public DeleteCommand(Ui ui, TaskList taskList, int index, Storage storage, File file) throws
             InvalidCmdValueException {
-        if (index + 1 > taskList.getSize() || index <= 0) {
+        if (index + 1 > taskList.getSize() || index < 0) {
             throw new InvalidCmdValueException();
         }
         this.ui = ui;
