+++ src/main/java/duke/exceptions/EmptyCommandException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class EmptyCommandException extends Exception {
 
     public EmptyCommandException(String taskType) {
