+++ src/main/java/duke/exceptions/InvalidTaskTypeException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidTaskTypeException extends Exception {
 
     public InvalidTaskTypeException() {
