+++ src/main/java/duke/exceptions/InvalidTimeException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidTimeException extends Exception {
     public InvalidTimeException() {
         super ("OOPS!!! Please specify a time.");
