+++ src/main/java/duke/tasks/Deadline.java
@@ -1,4 +1,7 @@
-import java.time.DateTimeException;
+package duke.tasks;
+
+import duke.exceptions.InvalidDateException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
