+++ src/main/java/duke/tasks/Event.java
@@ -1,3 +1,7 @@
+package duke.tasks;
+
+import duke.exceptions.InvalidDateException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
@@ -23,7 +27,7 @@ public class Event extends Task {
     @Override
     public String toString() {
         return "[E]" + super.toString() + "(from: " + start.format(formatter) +
-                " to:" + end.format(formatter) + ")";
+                " to: " + end.format(formatter) + ")";
     }
 
     @Override
