+++ src/main/java/duke/tasks/TaskTypes.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 public enum TaskTypes {
     TODO,
     DEADLINE,
