+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Duke
+Hello! I'm duke.Duke
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
